backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
redis==5.0.1
google-generativeai==0.3.2
tavily-python==0.3.1
praw==7.7.1
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0
python-dotenv==1.0.1
